# apps/web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "llm-observe API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/traces": {
      "get": {
        "tags": [
          "traces"
        ],
        "summary": "List traces, newest first",
        "description": "Return one page of traces, newest first, without their payloads.\n\nOrdered by `(started_at DESC, id DESC)`. The id is not decoration: traces\nwritten by one batch routinely share a timestamp, and it is the tiebreaker\nthat gives the sort a total order, which is the whole basis for the cursor\nbeing stable across pages.\n\nThe range filters apply to `started_at` - event time, when the run actually\nhappened - not to `created_at`, which is when this server received it. They\ndiffer whenever a client buffered, retried, or was offline.",
        "operationId": "list_traces_v1_traces_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Opaque cursor from a previous page.",
              "title": "Cursor"
            },
            "description": "Opaque cursor from a previous page."
          },
          {
            "name": "name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Exact trace name match.",
              "title": "Name"
            },
            "description": "Exact trace name match."
          },
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Exact match on the client's end-user id.",
              "title": "User Id"
            },
            "description": "Exact match on the client's end-user id."
          },
          {
            "name": "session_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Exact match on the client's session id.",
              "title": "Session Id"
            },
            "description": "Exact match on the client's session id."
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only traces with started_at >= this (inclusive).",
              "title": "From"
            },
            "description": "Only traces with started_at >= this (inclusive)."
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only traces with started_at < this (exclusive).",
              "title": "To"
            },
            "description": "Only traces with started_at < this (exclusive)."
          },
          {
            "name": "X-Project-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Project-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TraceListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Malformed cursor"
          },
          "401": {
            "description": "Missing, unrecognised, or revoked API key"
          },
          "403": {
            "description": "Key lacks the 'read' scope"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TraceListItem": {
        "properties": {
          "observation_count": {
            "type": "integer",
            "title": "Observation Count",
            "default": 0
          },
          "prompt_tokens": {
            "type": "integer",
            "title": "Prompt Tokens",
            "default": 0
          },
          "completion_tokens": {
            "type": "integer",
            "title": "Completion Tokens",
            "default": 0
          },
          "total_tokens": {
            "type": "integer",
            "title": "Total Tokens",
            "default": 0
          },
          "total_cost_usd": {
            "type": "string",
            "format": "decimal",
            "title": "Total Cost Usd",
            "default": "0",
            "examples": [
              "0.00123400"
            ]
          },
          "total_latency_ms": {
            "type": "integer",
            "title": "Total Latency Ms",
            "default": 0
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "ended_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "sdk_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sdk Version"
          },
          "duration_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Ms"
          }
        },
        "type": "object",
        "required": [
          "id",
          "metadata",
          "started_at",
          "created_at"
        ],
        "title": "TraceListItem",
        "description": "One row of GET /v1/traces."
      },
      "TraceListResponse": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/TraceListItem"
            },
            "type": "array",
            "title": "Data"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "TraceListResponse",
        "description": "A page of traces plus the cursor that continues it.\n\n`next_cursor` is opaque and must be echoed back verbatim. It encodes the\n(started_at, id) of the last row on this page, which is what makes paging\nstable when traces share a timestamp - id breaks the tie and the seek uses\nthe same pair the ORDER BY does."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "A project API key, or a dashboard session token from /v1/auth/google",
        "scheme": "bearer"
      }
    }
  }
}